Added the remove method
When removing the element is made in time complexity O(1) (constant time), I just replace the erased element by the last element of the array. But when copying the temporal new array to the main array this is made in O(n).

The resize() method increments by two ints sizes only if the current sizes is equal to the capacity of the main array.